add: user search without auth
